Remove sharedInstance and private init.
Making BluetoothUI a singleton doesn't really give any advantages.
On the flip side it introduces uncertainty about the state of the
handlers (anything can set them). Also, since they're not stored as a
property it can be difficult to see when singletons are a dependency,
so I'm making this a regularly instantiated class.